Introduced TiledIndex which wraps the original "uint" data of the layer data. This saves memory because the flags buffer is no longer needed.
