+++ data/exotic/Tiles32x32.A.tsx
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.5" tiledversion="1.7.2" name="Tiles A" tilewidth="32" tileheight="32" tilecount="3" columns="0">
+ <grid orientation="orthogonal" width="1" height="1"/>
+ <tile id="0">
+  <properties>
+   <property name="Text" value="Hello world!"/>
+  </properties>
+  <image width="32" height="32" source="tile1.png"/>
+  <objectgroup draworder="index" id="2">
+   <object id="5" x="4.25" y="2.125" width="23.375" height="28.625">
+    <ellipse/>
+   </object>
+  </objectgroup>
+ </tile>
+ <tile id="1">
+  <image width="32" height="32" source="tile1.png"/>
+ </tile>
+ <tile id="2">
+  <image width="32" height="32" source="tile1.png"/>
+ </tile>
+</tileset>
